requirements.txt
streamlit
torch
torchvision
Pillow
numpy
faiss-cpu
rsa
